moved each aggregator to separate file
